Implement `NoEq` variants of functions in Halley
Bump version number to 4.3.2 because of new functionality

These are variants which don't require an instance for Eq on the type
being operated on so they return an infinite list of results instead.

The existing functions have been modified to call the more general NoEq
variant and truncate the list returned from that with the new function
takeWhileDifferent.